Zilliz/requirements.txt
python-dotenv
pymilvus
sentence-transformers
numpy
fastapi
uvicorn
datasets
tqdm
streamlit
pandas
requests 